finished all controllers, everything working, needs styling
