+++ views/layout.ejs
@@ -16,6 +16,12 @@
             <li>
                 <a href="/dinosaurs/new">New Dino</a>
             </li>
+            <li>
+                <a href="/prehistoric_creatures/">Prehistoric Creatures Home</a>
+            </li>
+            <li>
+                <a href="/prehistoric_creatures/new">New Prehistoric Creatures</a>
+            </li>
         </ul>
     </nav>
     <h1>I LOVE DINOS!</h1>
