+++ views/prehistoric_creatures/index.ejs
@@ -1,0 +1,10 @@
+<ul>
+    <% creatureData.forEach((creature, index) => { %>
+        <li>
+            <%= creature.type %>
+            <img src="<%= creature.img_url %>" alt="">
+            <!-- NEED TO EDIT LINK -->
+            <a href="/prehistoric_creatures/edit/<%= index %>">Edit creature</a>
+        </li>
+    <% }) %>
+</ul>
