+++ views/prehistoric_creatures/new.ejs
@@ -1,0 +1,9 @@
+<form method="POST" action="/prehistoric_creatures/">
+    <label for="creatureType">Type: </label>
+    <input id="creatureType" type="text" name="type" />
+    
+    <label for="creatureImgUrl">Img URL: </label>
+    <input id="creatureImgUrl" type="text" name="img_url" />
+
+    <input type="submit" />
+</form>
